#5 extract BaseCallbackAnalyzer
