+++ Abc.MoqComplete/Abc.MoqComplete/CodeAnalysis/AutoMockerSuspiciousCallbackAnalyzer.cs
@@ -10,80 +10,20 @@ using JetBrains.ReSharper.Psi.Tree;
 namespace Abc.MoqComplete.CodeAnalysis
 {
 	[ElementProblemAnalyzer(typeof(IInvocationExpression), HighlightingTypes = new[] { typeof(AutoMockerSuspiciousCallbackWarning) })]
-	public class AutoMockerSuspiciousCallbackAnalyzer : ElementProblemAnalyzer<IInvocationExpression>
+	public class AutoMockerSuspiciousCallbackAnalyzer : BaseCallbackAnalyzer
 	{
 		/// <inheritdoc />
-		protected override void Run(IInvocationExpression element, ElementProblemAnalyzerData data, IHighlightingConsumer consumer)
+		protected override TreeNodeCollection<ICSharpArgument>? GetArguments(ISolution solution, IInvocationExpression methodInvocation)
 		{
-			var methodIdentifier = element.GetSolution().GetComponent<IMoqMethodIdentifier>();
-			var mockedMethodProvider = element.GetSolution().GetComponent<IAutoMockerMockedMethodProvider>();
+			var mockedMethodProvider = solution.GetComponent<IAutoMockerMockedMethodProvider>();
 
-			if (!methodIdentifier.IsMoqCallbackMethod(element))
-			{
-				return;
-			}
-
-			var expectedTypeParameters = element.TypeArguments;
-
-			if (expectedTypeParameters.Count == 0)
-			{
-				return;
-			}
-
-			var pointer = element.InvokedExpression;
-			TreeNodeCollection<ICSharpArgument>? arguments = null;
-
-			while (pointer != null && arguments == null && pointer.FirstChild is IInvocationExpression methodInvocation)
-			{
-				arguments = mockedMethodProvider.GetMockedMethodParametersFromSetupMethod(methodInvocation);
-				pointer = methodInvocation.InvokedExpression;
-			}
-
-			if (arguments == null)
-			{
-				return;
-			}
-
-			var actualTypesParameters = arguments.Value.Select(x => x.Value.Type()).ToArray();
-			var rule = element.GetPsiModule().GetTypeConversionRule();
-
-			if (actualTypesParameters.Length <= 0)
-			{
-				return;
-			}
-
-			if (expectedTypeParameters.Count != actualTypesParameters.Length)
-			{
-				AddWarning(element, consumer);
-			} else
-			{
-				for (var i = 0; i < expectedTypeParameters.Count; i++)
-				{
-					var actualParameterType = actualTypesParameters[i];
-					var expectedParameterType = expectedTypeParameters[i];
-
-					if (!actualParameterType.Equals(expectedParameterType)
-						&& !actualParameterType.IsImplicitlyConvertibleTo(expectedParameterType, rule))
-					{
-						AddWarning(element, consumer);
-					}
-				}
-			}
+			return mockedMethodProvider.GetMockedMethodParametersFromSetupMethod(methodInvocation);
 		}
 
-		private static void AddWarning(IInvocationExpression element, IHighlightingConsumer consumer)
+		/// <inheritdoc />
+		protected override void AddHighlighting(IHighlightingConsumer consumer, DocumentRange range)
 		{
-			DocumentRange range;
-
-			if (element.FirstChild?.LastChild is ITypeArgumentList typeInvocation)
-			{
-				range = typeInvocation.GetDocumentRange();
-			} else
-			{
-				range = element.InvokedExpression.GetDocumentRange();
-			}
-
-			consumer.AddHighlighting(new AutoMockerSuspiciousCallbackWarning(element, range));
+			consumer.AddHighlighting(new AutoMockerSuspiciousCallbackWarning(range));
 		}
 	}
 }
