+++ Abc.MoqComplete/Abc.MoqComplete/CodeAnalysis/AutoMockerSuspiciousCallbackWarning.cs
@@ -1,13 +1,12 @@
 ﻿using JetBrains.DocumentModel;
 using JetBrains.ReSharper.Feature.Services.Daemon;
-using JetBrains.ReSharper.Psi.CSharp.Tree;
 
 namespace Abc.MoqComplete.CodeAnalysis
 {
 	[StaticSeverityHighlighting(Severity.WARNING, HighlightingGroupIds.GutterMarksGroup)]
 	public sealed class AutoMockerSuspiciousCallbackWarning : BaseCallbackWarning
 	{
-		public AutoMockerSuspiciousCallbackWarning(IInvocationExpression element, DocumentRange documentRange)
+		public AutoMockerSuspiciousCallbackWarning(DocumentRange documentRange)
 			: base("AutoMocker suspicious Callback method call: Generic types do not match", documentRange)
 		{
 		}
